Moved all chatroom-related methods and state to RoomContext global state.
