+++ client/src/components/Chat/index.js
@@ -7,12 +7,10 @@ import axios from "axios";
 import './Chat.css';
 // Messages is provided as a props, so we need to retrieve it via props and then destructure messages out of props
 function Chat(props) {
-    const roomContext = useContext(RoomContext);
-    const { messages, chatroomName } = props;
-    const [input, setInput] = useState("");
     const authContext = useContext(AuthContext);
+    const roomContext = useContext(RoomContext);
 
-    // const lastSeen = messages[messages.length - 1].timeStamp;
+    const [input, setInput] = useState("");
 
     const sendMessage = async (event) => {
         event.preventDefault();
@@ -26,7 +24,7 @@ function Chat(props) {
             chatroomID: roomContext.currentRoomID
         });
         //Now, run parent function to get the messages again and render messages again.
-        props.handlePusher();
+        roomContext.handlePusher();
         //props.getMessages();
         //Once axios has completed, set the input back to blank
         setInput("");
@@ -38,15 +36,14 @@ function Chat(props) {
             <div className="chat_header">
                 <Avatar />
                 <div className='chat_headerInfo'>
-                    <h3>{chatroomName}</h3>
-                    <p>Last message sent: </p>
+                    <h3>{roomContext.currentRoomName}</h3>
                 </div>
             </div>
 
             <div className='chat_body' id='chat_body'>
                 {/* Here we loop through messages and create a new chat bubble for each message. if the .recieved is true,
                  the bubble will be given the className 'chat_reciever' for different styling*/}
-                {messages.map((message, index) => (
+                {roomContext.messages.map((message, index) => (
                     <>
                         <p key={index} className={`chat_message ${message.senderID === authContext.user._id ? "chat_reciever" : ""}`}>
                         <span className='chat_name'>{message.name}</span>
